pyproject.toml
[project]
name = "agent-design-pattern"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "groq>=0.29.0",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.1",
]
